+++ noQt/shaders/simulation/subtractGradient.frag
@@ -11,7 +11,7 @@ layout (binding = 2) uniform sampler3D obstacleTexture;
 
 layout (location = 3) uniform float gridScale;
 
-flat in int layer;
+flat in float layer;
 
 void main()
 {
